+++ src/MetaheuristicsPlatform/Algorithms/Multiobjective/MoeadDe/MoeadDeReferences.cs
@@ -1,0 +1,12 @@
+using MetaheuristicsPlatform.Classification;
+namespace MetaheuristicsPlatform.Algorithms.Multiobjective.MoeadDe;
+public static class MoeadDeReferences
+{
+    public static ScientificReference LiZhang2009 { get; } =
+        new(
+            "Li & Zhang",
+            2009,
+            "Multiobjective Optimization Problems With Complicated Pareto Sets, MOEA/D and NSGA-II",
+            "IEEE Transactions on Evolutionary Computation 13(2), 284-302",
+            "10.1109/TEVC.2008.925798");
+}